Refactor account info to split client side and server side rendering
